+++ VpnHood.Client/SessionStatus.cs
@@ -1,13 +1,14 @@
-﻿using VpnHood.Tunneling.Messages;
+﻿using VpnHood.Common.Messaging;
+using VpnHood.Tunneling.Messaging;
 
 namespace VpnHood.Client
 {
     public class SessionStatus
     {
-        public ResponseCode ResponseCode { get; internal set; }
-        public AccessUsage AccessUsage { get; internal set; } = new AccessUsage();
-        public SuppressType SuppressedTo { get; internal set; }
-        public SuppressType SuppressedBy { get; internal set; }
+        public SessionErrorCode ErrorCode { get; internal set; } //todo change UI from response coce
+        public AccessUsage? AccessUsage { get; internal set; } 
+        public SessionSuppressType SuppressedTo { get; internal set; }
+        public SessionSuppressType SuppressedBy { get; internal set; }
         public string? ErrorMessage { get; internal set; }
     }
 }