+++ VpnHood.Common/Messaging/ClientInfo.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace VpnHood.Common.Messaging
+{
+    public class ClientInfo
+    {
+        public Guid ClientId { get; set; }
+        public string? UserToken { get; set; }
+        public string ClientVersion { get; set; } = null!;
+        public int ProtocolVersion { get; set; }
+        public string UserAgent { get; set; } = null!;
+
+        public override string ToString() 
+            => $"{nameof(ClientId)}={ClientId}";
+    }
+}