+++ VpnHood.Tunneling/Messages/UdpChannelRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace VpnHood.Tunneling.Messages
-{
-    public class UdpChannelRequest : SessionRequest
-    {
-
-    }
-}