+++ VpnHood.Tunneling/Messaging/SessionException.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Net;
 using VpnHood.Common.Messaging;
 
-namespace VpnHood.Server
+namespace VpnHood.Tunneling.Messaging
 {
     public class SessionException : Exception
     {