+++ VpnHood.Tunneling/TcpClientStream.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
 using System.IO;
+using System.Net;
 using System.Net.Sockets;
 using System.Threading;
 using VpnHood.Logging;
@@ -12,6 +13,8 @@ namespace VpnHood.Tunneling
         private bool _disposed = false;
         public TcpClient TcpClient { get; }
         public Stream Stream { get; set; }
+        public IPEndPoint LocalEndPoint => (IPEndPoint)TcpClient.Client.LocalEndPoint;
+        public IPEndPoint RemoteEndPoint => (IPEndPoint)TcpClient.Client.RemoteEndPoint;
 
         public TcpClientStream(TcpClient tcpClient, Stream stream)
         {